Improve the module's structure
Refactor the match the structure of `src/sorting`.
